Remove account_utils duplication
